user model completed

+++ src/main/java/models/Bookshelf.java
@@ -8,7 +8,7 @@ public class Bookshelf { // K & R
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn (name = "user_id")
@@ -36,12 +36,12 @@ public class Bookshelf { // K & R
 
     public Bookshelf(){}
 
-    public int getId(){
+    public Long getId(){
         return id;
     }
 
     public
-    void setId(int id){
+    void setId(Long id){
         this.id = id;
     }
